Moved programming exercise specs into their own folder
