+++ src/test/cypress/integration/exercises/programming/ProgrammingExerciseParticipation.spec.ts
@@ -1,9 +1,9 @@
-import { CypressCredentials } from '../support/users';
-import { generateUUID } from '../support/utils';
-import allSuccessful from '../fixtures/programming_exercise_submissions/all_successful/submission.json';
-import partiallySuccessful from '../fixtures/programming_exercise_submissions/partially_successful/submission.json';
-import { artemis } from '../support/ArtemisTesting';
-import { ProgrammingExerciseSubmission } from '../support/pageobjects/OnlineEditorPage';
+import { CypressCredentials } from '../../../support/users';
+import { generateUUID } from '../../../support/utils';
+import allSuccessful from '../../../fixtures/programming_exercise_submissions/all_successful/submission.json';
+import partiallySuccessful from '../../../fixtures/programming_exercise_submissions/partially_successful/submission.json';
+import { artemis } from '../../../support/ArtemisTesting';
+import { ProgrammingExerciseSubmission } from '../../../support/pageobjects/OnlineEditorPage';
 
 // The user management object
 const users = artemis.users;
